Improved exception handling and changed how errors are handled to Python land

+++ Connection.cpp
@@ -84,6 +84,7 @@ Connection::Connection (AMConnectionCAPI *_capi)
 	m_errno = -1;
 	memcpy (&m_capi, _capi, sizeof (AMConnectionCAPI));
 	m_dbgMethodProgress = 0;
+	m_errorType = AME_OTHER;
 }
 
 Connection::~Connection()
@@ -141,7 +142,7 @@ bool Connection::readSocket()
 	{
 		// Socket buffer got full!
 
-		setError("Socket receive buffer full", 0);
+		setError("Socket receive buffer full", 0, AME_OTHER);
 		return false;
 	}
 
@@ -155,13 +156,13 @@ bool Connection::readSocket()
 		}
 
 		int sockError = SocketGetLastError();
-		setError("Socket receive failed", sockError);
+		setError("Socket receive failed",  sockError, AME_OTHER);
 		return false;
 	}
 	else
 	if (recvResult == 0)
 	{
-		setError("Connection reset by peer when receiving", 0);
+		setError("Connection reset by peer when receiving", 0, AME_OTHER);
 		return false;
 	}
 
@@ -186,7 +187,7 @@ bool Connection::writeSocket()
 			return true;
 		}
 
-		setError("Socket send failed", SocketGetLastError());
+		setError("Socket send failed", SocketGetLastError(), AME_OTHER);
 		return false;
 	}
 
@@ -238,13 +239,13 @@ bool Connection::processHandshake()
 	
 	if (protocolVersion == 0xff)
 	{
-		setError("Too many connections reported by server", 0);
+		setError("Too many connections reported by server", 0, AME_OTHER);
 		return false;
 	}
 	else
 	if (protocolVersion != MYSQL_PROTOCOL_VERSION)
 	{
-		setError("Protocol version not supported(1)", 0);
+		setError("Protocol version not supported(1)", 0, AME_OTHER);
 		return false;
 	}
 
@@ -257,7 +258,7 @@ bool Connection::processHandshake()
 
 	if (!(serverCaps & MCP_PROTOCOL_41))
 	{
-		setError("Authentication < 4.1 not supported", 1);
+		setError("Authentication < 4.1 not supported", 1, AME_OTHER);
 		return false;
 	}
 
@@ -273,7 +274,7 @@ bool Connection::processHandshake()
 	}
 	else
 	{
-		setError("Authentication < 4.1 not supported", 2);
+		setError("Authentication < 4.1 not supported", 2, AME_OTHER);
 		return false;
 	}
 
@@ -284,7 +285,7 @@ bool Connection::processHandshake()
 
 	if (!(serverCaps & MCP_CONNECT_WITH_DB) && !m_database.empty())
 	{
-		setError("Protocol < 4.1 not supported", 3);
+		setError("Protocol < 4.1 not supported", 3, AME_OTHER);
 		return false;
 	}
 
@@ -395,24 +396,30 @@ bool Connection::sendPacket()
 	return true;
 }
 
-void Connection::setError (const char *_message, int _errno)
+void Connection::setError (const char *_message, int _errno, AMErrorType _type)
 {
 	m_errorMessage = _message;
 	m_errno = _errno;
+	m_errorType = _type;
 
 	PRINTMARK();
 
-	if (m_sockInst)
+
+	if (_type != AME_MYSQL)
 	{
-		PRINTMARK();
-		m_capi.closeSocket(m_sockInst);
-		m_capi.deleteSocket(m_sockInst);
-		m_sockInst = NULL;
+
+		if (m_sockInst)
+		{
+			PRINTMARK();
+			m_capi.closeSocket(m_sockInst);
+			m_capi.deleteSocket(m_sockInst);
+			m_sockInst = NULL;
+		}
 	}
 }
 
 
-bool Connection::getLastError (const char **_ppOutMessage, int *_outErrno)
+bool Connection::getLastError (const char **_ppOutMessage, int *_outErrno, int *_outErrorType)
 {
 	if (m_errno == -1)
 	{
@@ -420,6 +427,7 @@ bool Connection::getLastError (const char **_ppOutMessage, int *_outErrno)
 	}
 
 	*_ppOutMessage = m_errorMessage.c_str();
+	*_outErrorType = (int) m_errorType;
 	*_outErrno = m_errno;
 
 	m_errno = -1;
@@ -434,11 +442,11 @@ bool Connection::connect(const char *_host, int _port, const char *_username, co
 
 	if (m_dbgMethodProgress > 1)
 	{
-		fprintf (stderr, "%s:%d: UNEXPECTED:>\n", __FUNCTION__, __LINE__);
 		/*
 		NOTE: We don't call setError here because it will close the socket worsening the concurrent access error making it impossible to trace */
 		m_errorMessage = "Concurrent access in connect method";
 		m_errno = 0;
+		m_errorType = AME_OTHER;
 		m_dbgMethodProgress --;
 		return false;
 	}
@@ -447,7 +455,7 @@ bool Connection::connect(const char *_host, int _port, const char *_username, co
 	if (m_sockInst != NULL)
 	{
 		m_dbgMethodProgress --;
-		setError ("Socket already connected", 0);
+		setError ("Socket already connected", 0, AME_OTHER);
 		return false;
 	}
 
@@ -473,7 +481,6 @@ bool Connection::connect(const char *_host, int _port, const char *_username, co
 		if (!setTimeout (m_timeout))
 		{
 			m_dbgMethodProgress --;
-			setError("setTimeout API failed", 0);
 			return false;
 		}
 	}
@@ -588,7 +595,7 @@ void Connection::handleErrorPacket()
 	UINT8 *message = m_reader.readBytes(len);
 
 	std::string errorMessage((char *) message, len);
-	setError (errorMessage.c_str (), (int) errnum);
+	setError (errorMessage.c_str (), (int) errnum, AME_MYSQL);
 }
 
 void *Connection::handleResultPacket(int _fieldCount)
@@ -716,12 +723,11 @@ void *Connection::query(const char *_query, size_t _cbQuery)
 
 	if (m_dbgMethodProgress > 1)
 	{
-		fprintf (stderr, "%s:%d: UNEXPECTED:>\n", __FUNCTION__, __LINE__);
-
 		/*
 		NOTE: We don't call setError here because it will close the socket worsening the concurrent access error making it impossible to trace */
 		m_errorMessage = "Concurrent access in query method";
 		m_errno = 0;
+		m_errorType = AME_OTHER;
 		m_dbgMethodProgress --;
 		return NULL;
 	}
@@ -729,7 +735,7 @@ void *Connection::query(const char *_query, size_t _cbQuery)
 	if (m_sockInst == NULL)
 	{
 		PRINTMARK();
-		setError ("Not connected", 0);
+		setError ("Not connected", 0, AME_OTHER);
 		m_dbgMethodProgress --;
 		return NULL;
 	}
@@ -739,7 +745,7 @@ void *Connection::query(const char *_query, size_t _cbQuery)
 	if (len > m_writer.getSize () - (MYSQL_PACKET_HEADER_SIZE + 1))
 	{
 		PRINTMARK();
-		setError ("Query too big", 0);
+		setError ("Query too big", 0, AME_OTHER);
 		m_dbgMethodProgress --;
 		return NULL;
 	}
@@ -780,7 +786,7 @@ void *Connection::query(const char *_query, size_t _cbQuery)
 
 		case 0xfe:
 			PRINTMARK();
-			setError ("Unexpected EOF when decoding result", 0);
+			setError ("Unexpected EOF when decoding result", 0, AME_OTHER);
 			m_dbgMethodProgress --;
 			return NULL;
